[refactor] xoá modal-overlay-avatar (vì đã có phương án thay thế)
