+++ src/main/java/filter/WebFilter.java
@@ -1,4 +1,4 @@
-package controller;
+package filter;
 
 import jakarta.servlet.*;
 import jakarta.servlet.http.*;
