+++ src/main/webapp/admin/manage_banner.jsp
@@ -155,22 +155,6 @@
             </div>
         </main>
     </div>
-    <%@ include file="/admin/components/notify_modal.jsp" %>
-    <div class="modal-overlay-avatar" id="avatar-account-modal">
-        <div class="modal-content-avatar">
-            <button class="modal-close2" id="close-modal-btn9">
-                <ion-icon name="close-outline"></ion-icon>
-            </button>
-            <a href="${pageContext.request.contextPath}/home" class="btn-menu-item">
-                <ion-icon name="person-circle-outline"></ion-icon>
-                <span>Trở về trang người dùng</span>
-            </a>
-            <a href="${pageContext.request.contextPath}/logout" class="btn-menu-item">
-                <ion-icon name="log-out-outline"></ion-icon>
-                <span>Đăng xuất tài khoản</span>
-            </a>
-        </div>
-    </div>
 </div>
 
 <div class="modal-overlay-form product-form-modal" id="banner-form-modal">
