+++ src/main/webapp/admin/manage_blog.jsp
@@ -201,23 +201,6 @@
                     </div>
                 </div>
 
-                <%@ include file="/admin/components/notify_modal.jsp" %>
-                    <div class="modal-overlay-avatar" id="avatar-account-modal">
-                        <div class="modal-content-avatar">
-                            <button class="modal-close2" id="close-modal-btn9">
-                                <ion-icon name="close-outline"></ion-icon>
-                            </button>
-                            <a href="${pageContext.request.contextPath}/home" class="btn-menu-item">
-                                <ion-icon name="person-circle-outline"></ion-icon>
-                                <span>Trở về trang người dùng</span>
-                            </a>
-                            <a href="${pageContext.request.contextPath}/logout" class="btn-menu-item">
-                                <ion-icon name="log-out-outline"></ion-icon>
-                                <span>Đăng xuất tài khoản</span>
-                            </a>
-                        </div>
-                    </div>
-
                     <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
                     <link rel="stylesheet" href="https://cdn.datatables.net/2.3.4/css/dataTables.dataTables.css" />
                     <script src="https://cdn.datatables.net/2.3.4/js/dataTables.js"></script>
