+++ src/test/java/WebFilterTest.java
@@ -1,4 +1,4 @@
-import controller.WebFilter;
+import filter.WebFilter;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
